+++ vite.config.ts
@@ -51,6 +51,7 @@ export default defineConfig({
       '@gluonjs/ssr/streaming': resolve(import.meta.dirname, 'packages/ssr/src/streaming.ts'),
       '@gluonjs/ssr': resolve(import.meta.dirname, 'packages/ssr/src/index.ts'),
       '@gluonjs/store': resolve(import.meta.dirname, 'packages/store/src/index.ts'),
+      '@gluonjs/test-utils/ssr': resolve(import.meta.dirname, 'packages/test-utils/src/ssr.ts'),
       '@gluonjs/test-utils': resolve(import.meta.dirname, 'packages/test-utils/src/index.ts'),
     },
   },